requirements.txt
argcomplete==3.1.1
astroid==2.15.6
asttokens==2.2.1
backcall==0.2.0
black==23.7.0
blessed==1.20.0
cachetools==5.3.1
certifi==2023.7.22
cfgv==3.3.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.1
colorama==0.4.6
comm==0.1.4
commitizen==3.6.0
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decli==0.6.1
decorator==5.1.1
dill==0.3.7
distlib==0.3.8
executing==1.2.0
filelock==3.12.2
flake8==6.1.0
fonttools==4.42.0
gpustat==1.1.1
identify==2.5.26
idna==3.4
importlib-metadata==6.8.0
imutils==0.5.4
ipykernel==6.25.1
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lightning-utilities==0.9.0
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mpmath==1.3.0
mypy==1.4.1
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
nodeenv==1.8.0
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-contrib-python==4.9.0.80
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pep8-naming==0.13.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==4.2.2
pluggy==1.2.0
pre-commit==3.3.3
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.16.1
pylint==2.17.5
pyparsing==3.0.9
pyproject-api==1.5.3
python-dateutil==2.8.2
pytz==2023.3
pyupgrade==3.10.1
PyYAML==6.0.1
pyzmq==25.1.0
questionary==1.10.0
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
stack-data==0.6.2
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.2.0
tokenize-rt==5.2.0
tomli==2.0.1
tomlkit==0.12.1
torch==2.0.0
torch-frft @ git+https://github.com/tunakasif/torch-frft.git@4dbfce77e2e3fef2cb54e7aa66a4bbb3037b7ff2
torchmetrics==1.1.0
torchvision==0.15.2
tornado==6.3.2
tox==4.6.4
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.26.2
wcwidth==0.2.6
wrapt==1.15.0
zipp==3.16.2
